# openapi.json
{"openapi":"3.1.0","info":{"title":"tree-health-infer-service","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/infer":{"post":{"tags":["inference"],"summary":"Infer","operationId":"infer_infer_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_infer_infer_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InferResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/debug/depth":{"post":{"tags":["debug"],"summary":"Depth","operationId":"depth_debug_depth_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_depth_debug_depth_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/debug/heuristics":{"post":{"tags":["debug"],"summary":"Heuristics","operationId":"heuristics_debug_heuristics_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_heuristics_debug_heuristics_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/debug/export-smoke":{"post":{"tags":["debug"],"summary":"Export Smoke","operationId":"export_smoke_debug_export_smoke_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_export_smoke_debug_export_smoke_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BBox":{"properties":{"x1":{"type":"integer","title":"X1"},"y1":{"type":"integer","title":"Y1"},"x2":{"type":"integer","title":"X2"},"y2":{"type":"integer","title":"Y2"}},"type":"object","required":["x1","y1","x2","y2"],"title":"BBox"},"Body_depth_debug_depth_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_depth_debug_depth_post"},"Body_export_smoke_debug_export_smoke_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_export_smoke_debug_export_smoke_post"},"Body_heuristics_debug_heuristics_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_heuristics_debug_heuristics_post"},"Body_infer_infer_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_infer_infer_post"},"DefectItem":{"properties":{"id":{"type":"integer","title":"Id"},"cls":{"type":"string","title":"Cls"},"conf":{"type":"number","title":"Conf"},"bbox":{"$ref":"#/components/schemas/BBox"},"area":{"type":"integer","title":"Area"},"plant_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Plant Id"},"mask_rle":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mask Rle"},"area_ratio":{"type":"number","title":"Area Ratio","default":0.0},"severity":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Severity"}},"type":"object","required":["id","cls","conf","bbox","area"],"title":"DefectItem"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InferResponse":{"properties":{"request_id":{"type":"string","title":"Request Id"},"status":{"type":"string","title":"Status","description":"OK | NO_PLANTS | ERROR"},"overlay_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Overlay Path"},"report_json_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Report Json Path"},"plants":{"items":{"$ref":"#/components/schemas/PlantItem"},"type":"array","title":"Plants","default":[]},"defects":{"items":{"$ref":"#/components/schemas/DefectItem"},"type":"array","title":"Defects","default":[]},"extras":{"additionalProperties":{"type":"string"},"type":"object","title":"Extras","default":{}}},"type":"object","required":["request_id","status"],"title":"InferResponse"},"PlantItem":{"properties":{"id":{"type":"integer","title":"Id"},"cls":{"type":"string","title":"Cls"},"conf":{"type":"number","title":"Conf"},"bbox":{"$ref":"#/components/schemas/BBox"},"area":{"type":"integer","title":"Area"},"tilt_deg":{"type":"number","title":"Tilt Deg"},"dry_ratio":{"type":"number","title":"Dry Ratio"},"species":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Species"},"mask_rle":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mask Rle"},"health_score":{"type":"number","title":"Health Score","default":100.0},"health_grade":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Health Grade"}},"type":"object","required":["id","cls","conf","bbox","area","tilt_deg","dry_ratio"],"title":"PlantItem"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}